#9: Add README and reorganize library
