+++ README.md
@@ -1,2 +1,35 @@
-# reso-certification-etl
-Certification ETL Package
+# RESO Certification Common Library
+NodeJS Package Containing RESO Certification ETL utilities.
+
+# Installation
+To install from GitHub using npm:
+
+```
+npm i RESOStandards/reso-certification-common
+```
+
+To install from Github via yarn:
+```
+yarn add https://github.com/RESOStandards/reso-certification-etl.git
+```
+
+## Usage
+See `index.js` for list of exported libraries.
+
+
+## Tests
+To run the tests, clone the repository:
+```
+git clone https://github.com/RESOStandards/reso-certification-etl.git
+```
+
+Then change into the directory and run:
+```
+npm i
+```
+
+Finally:
+
+```
+npm test
+```
