+++ lib/process-lookup-resource-metadata.js
@@ -1,3 +1,45 @@
+const fs = require("fs/promises");
+const { Buffer } = require("node:buffer");
+
+const writeFile = async (path, content = "") => {
+  if (!path?.length) throw new Error("Path was empty in writeFile!");
+
+  try {
+    await fs.appendFile(path, Buffer.from(content));
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+const readFile = async (path = "") => {
+  if (!path?.length) throw new Error("Path was empty in readFile!");
+
+  try {
+    return await fs.readFile(path, { encoding: "utf8" });
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+const processLookupResourceMetadataFiles = async (
+  pathToMetadataReportJson = "",
+  pathToLookupResourceData = "",
+  pathToOutputFile = ""
+) => {
+  const metadataReportJson = JSON.parse(
+      await readFile(pathToMetadataReportJson)
+    ),
+    lookupResourceJson = JSON.parse(await readFile(pathToLookupResourceData));
+
+  const results = processLookupResourceMetadata({
+    metadataReportJson,
+    lookupResourceJson,
+  });
+
+  await writeFile(pathToOutputFile, JSON.stringify(results));
+};
+
+
 const EDM_STRING_TYPE = "Edm.String",
   LOOKUP_NAME_ANNOTATION_TERM = "RESO.OData.Metadata.LookupName",
   LOOKUP_STANDARD_NAME_ANNOTATION_TERM = "RESO.OData.Metadata.StandardName",
@@ -133,4 +175,5 @@ const processLookupResourceMetadata = ({
 
 module.exports = {
   processLookupResourceMetadata,
+  processLookupResourceMetadataFiles
 };
